+++ src/main/java/ru/planetnails/partnerslk/controllers/authentication/AuthenticationRequestDto.java
@@ -1,0 +1,15 @@
+package ru.planetnails.partnerslk.controllers.authentication;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticationRequestDto {
+    private String username;
+    private String password;
+}
