--- src/main/java/ru/planetnails/partnerslk/dto/AuthenticationRequestDto.java
@@ -1,12 +1,0 @@
-package ru.planetnails.partnerslk.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationRequestDto {
-    private String username;
-    private String password;
-}
