+++ src/main/java/ru/planetnails/partnerslk/repository/RoleRepository.java
@@ -1,7 +1,7 @@
 package ru.planetnails.partnerslk.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.planetnails.partnerslk.model.baseClasses.Role;
+import ru.planetnails.partnerslk.model.role.Role;
 
 public interface RoleRepository extends JpaRepository<Role, Long> {
     Role findByName(String name);
