+++ src/main/java/ru/planetnails/partnerslk/security/JwtUserDetailsService.java
@@ -1,7 +1,6 @@
 package ru.planetnails.partnerslk.security;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
