+++ src/main/java/ru/planetnails/partnerslk/security/jwt/JwtTokenProvider.java
@@ -1,20 +1,18 @@
 package ru.planetnails.partnerslk.security.jwt;
 
 import io.jsonwebtoken.*;
-import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.io.Encoders;
 import io.jsonwebtoken.security.Keys;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Component;
-import ru.planetnails.partnerslk.model.baseClasses.Role;
+import ru.planetnails.partnerslk.model.role.Role;
 
 import javax.annotation.PostConstruct;
 import javax.crypto.SecretKey;
